+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+
+### Removed
+
+## [1.23.1](https://pypi.org/project/model-bakery/1.23.1/)
+
+### Changed
 - Fix `_bulk_create` for ManyToManyField using custom `through` models ([#477](https://github.com/model-bakers/model_bakery/issues/477))
 - [dev] Clean up stale TODO and fix signal handler leak in m2m test